TST Move slow compile tests to nightly CI
These tests use torch.compile on semi-realistic models and are thus slow
to execute. In sum, they take ~3 min to finish, with an overall CI
runtime of ~15 min, so it's significant.

As these tests are very unlikley to be affected by most code changes, it
should be fine to move them to nightly CI instead of running them on
each PR. Also, presence of GPUs might speed the tests up.
